requirements.txt
flask
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
langchain-chroma
langchain-text-splitters
chromadb
sentence-transformers
beautifulsoup4
requests
pydantic
python-multipart
python-dotenv
unstructured
markdown
